requirements.txt
praw
schedule 
requests
python-dotenv
groq
markdown